Method implemented to check if HR_Database.txt file exists.

+++ src/project/Database/HR_Database/HRDatabaseController.java
@@ -1,5 +1,6 @@
 package project.Database.HR_Database;
 
+import javafx.scene.control.Alert;
 import project.Database.User.PersonalDetail;
 
 import java.io.*;
@@ -33,28 +34,48 @@ public class HRDatabaseController{
         return hrDatabase;
     }
 
+    /**
+     * Checks if HR_Database.txt File exists
+     * @return
+     */
+    public boolean checkIfHR_DatabaseFileExists()
+    {
+        File hr_DatabaseFile = new File(HR_DATABASE_FILE_PATH);
+        if(hr_DatabaseFile.exists() && !hr_DatabaseFile.isDirectory())
+        {
+            return true;
+        }else
+        {
+            showHR_DatabaseFileDoesNotExistAlert();
+            System.exit(0);
+            return false;
+        }
+    }
 
     /**
      * Reads the HR_Database.txt file and populates the hrDatabase ArrayList with its contents.
      */
     public void populateHRDatabase()
     {
-        FileInputStream fileInputStream;
-        ObjectInputStream objectInputStream;
 
-        try {
-            fileInputStream = new FileInputStream(HR_DATABASE_FILE_PATH);
-            objectInputStream = new ObjectInputStream(fileInputStream);
+            FileInputStream fileInputStream;
+            ObjectInputStream objectInputStream;
 
-            HRDatabase readHR_Database = (HRDatabase) objectInputStream.readObject();
-            hrDatabase = readHR_Database;
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
-        } catch (ClassNotFoundException e) {
-            e.printStackTrace();
-        }
+            try {
+                fileInputStream = new FileInputStream(HR_DATABASE_FILE_PATH);
+                objectInputStream = new ObjectInputStream(fileInputStream);
+
+                HRDatabase readHR_Database = (HRDatabase) objectInputStream.readObject();
+                hrDatabase = readHR_Database;
+            } catch (FileNotFoundException e) {
+                e.printStackTrace();
+            } catch (IOException e) {
+                e.printStackTrace();
+            } catch (ClassNotFoundException e) {
+                e.printStackTrace();
+            }
+
+
 
 
     }
@@ -64,18 +85,25 @@ public class HRDatabaseController{
      */
     public void writeToDatabase()
     {
-        FileOutputStream fileOutputStream;
-        ObjectOutputStream objectOutputStream;
-        try {
-            fileOutputStream = new FileOutputStream(HR_DATABASE_FILE_PATH);
-            objectOutputStream = new ObjectOutputStream(fileOutputStream);
+        if(checkIfHR_DatabaseFileExists())
+        {
+            FileOutputStream fileOutputStream;
+            ObjectOutputStream objectOutputStream;
+            try {
+                fileOutputStream = new FileOutputStream(HR_DATABASE_FILE_PATH);
+                objectOutputStream = new ObjectOutputStream(fileOutputStream);
 
-            objectOutputStream.writeObject(hrDatabase);
-        } catch (FileNotFoundException e) {
-            e.printStackTrace();
-        } catch (IOException e) {
-            e.printStackTrace();
+                objectOutputStream.writeObject(hrDatabase);
+            } catch (FileNotFoundException e) {
+                e.printStackTrace();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+        }else
+        {
+
         }
+
     }
 
     /**
@@ -110,5 +138,16 @@ public class HRDatabaseController{
         }
 
         return personalDetailReturn;
+    }
+
+    /**
+     * Shows Alert Box if HR_Database file does not exist
+     */
+    public void showHR_DatabaseFileDoesNotExistAlert()
+    {
+
+        Alert hr_DatabaseFileDoesNotExist = new Alert(javafx.scene.control.Alert.AlertType.ERROR);
+        hr_DatabaseFileDoesNotExist.setHeaderText(HR_DATABASE_FILE_PATH + " does not exist, application exiting");
+        hr_DatabaseFileDoesNotExist.show();
     }
 }
